Extract libs/household-analytics with alembic migrations as package data
Move the analytics contract and persistence layer into a new
policyengine-household-analytics workspace member: the calculate
analytics event schema, persistence, database setup, SQLAlchemy ORM
(data/models.py becomes orm.py), and the Alembic migration scripts,
which now ship inside the package. Revision ancestry checks resolve
scripts from the installed lib in every runtime — previously they
resolved from the repo root and silently fell back to string comparison
in images that shipped no alembic/ directory.

The core package depends on the analytics lib and drops its direct
alembic/flask-sqlalchemy/cloud-sql-connector/pymysql pins (still
supplied transitively). The debug-mode sqlite path no longer imports
the core package, keeping the analytics closure core-free. Worker
images stop copying alembic from the repo root; the alembic release
compatibility check and root alembic.ini point at the packaged scripts.

Part of #1605
