+++ alembic.ini
@@ -1,5 +1,5 @@
 [alembic]
-script_location = alembic
+script_location = policyengine_household_analytics:alembic
 prepend_sys_path = .
 path_separator = os
 sqlalchemy.url = sqlite://
