+++ gcp/cloud_run/worker.Dockerfile
@@ -9,8 +9,6 @@ RUN apt-get update && \
 RUN pip install --no-cache-dir -U uv
 
 COPY ./pyproject.toml ./uv.lock /build/
-COPY ./alembic.ini /build/alembic.ini
-COPY ./alembic /build/alembic
 COPY ./config /build/config
 COPY ./libs /build/libs
 
@@ -46,11 +44,10 @@ RUN apt-get update && \
 
 COPY --from=builder /opt/venv /opt/venv
 COPY --from=builder /build/libs/household-api/pyproject.toml /app/pyproject.toml
-COPY --from=builder /build/alembic.ini /app/alembic.ini
-COPY --from=builder /build/alembic /app/alembic
 COPY --from=builder /build/config /app/config
 COPY --from=builder /build/libs/household-api/policyengine_household_api /app/policyengine_household_api
 COPY --from=builder /build/libs/household-common/policyengine_household_common /app/policyengine_household_common
+COPY --from=builder /build/libs/household-analytics/policyengine_household_analytics /app/policyengine_household_analytics
 COPY ./gcp/cloud_run/worker_start.sh /app/start.sh
 RUN chmod +x /app/start.sh
