+++ libs/household-analytics/pyproject.toml
@@ -1,0 +1,32 @@
+[project]
+name = "policyengine-household-analytics"
+version = "0.26.1"
+description = "Analytics contract and persistence for the PolicyEngine Household API"
+readme = "README.md"
+authors = [
+    { name = "PolicyEngine", email = "hello@policyengine.org" },
+]
+requires-python = ">=3.12"
+# The slim analytics writer image installs this lib: keep numpy, country
+# model packages, and modal out of this closure.
+dependencies = [
+    "policyengine-household-common==0.26.1",
+    "alembic>=1.13.0",
+    "cloud-sql-python-connector",
+    "flask>=2.2",
+    "flask-sqlalchemy>=3",
+    "pymysql",
+    "sqlalchemy>=2",
+]
+
+[tool.uv.sources]
+policyengine-household-common = { workspace = true }
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+# Includes the alembic/ migration scripts as package data, so revision
+# ancestry checks work in every runtime that installs this lib.
+packages = ["policyengine_household_analytics"]
