+++ tests/fixtures/endpoints/analytics.py
@@ -3,8 +3,8 @@ from datetime import datetime
 import pytest
 from flask import Flask, Response
 
-from policyengine_household_api.data.analytics_setup import db
-from policyengine_household_api.data.models import (
+from policyengine_household_analytics.analytics_setup import db
+from policyengine_household_analytics.orm import (
     CalculateRequest,
     CalculateRequestVariable,
     Visit,
