+++ tests/fixtures/endpoints/household.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 import pytest
 
-from policyengine_household_api.analytics.events import CalculateAnalyticsEvent
+from policyengine_household_analytics.events import CalculateAnalyticsEvent
 from policyengine_household_common.models.analytics import (
     AnalyticsContext,
     AvailabilityStatus,
